% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\docType{package}
\name{metricsgraphics-package}
\alias{metricsgraphics-package}
\title{metrics-graphics.js for R}
\description{
metrics-graphics.js for R
}
\seealso{
\url{http://metricsgraphicsjs.org/}
}

