% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MG.R
\name{MG}
\alias{MG}
\alias{MG.data.frame}
\title{create a new MG plot object}
\usage{
MG(data = NULL, ...)

\method{MG}{data.frame}(data, missing_is_zero = NULL, x_accessor = NULL,
  y_accessor = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{data that at least has columns of \code{x_accessor} and \code{y_accessor}}

\item{missing_is_zero}{if true and if the data object is a time series, missing data points will be treated as zeros.}

\item{x_accessor}{the name of the element in the data that should be considered the x-accessor}

\item{y_accessor}{the name of the element in the data that should be considered the y-accessor}

\item{width}{the width of the output widget}

\item{height}{the height of the output widget}
}
\description{
Specify data options and create a MG plot object.
}
\seealso{
\url{https://github.com/mozilla/metrics-graphics/wiki/Data}
}

