% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MG_graphic.R
\name{MG_graphic}
\alias{MG_graphic}
\title{Add graphic options to a MG plot object}
\usage{
MG_graphic(animate_on_load = NULL, area = NULL, baselines = NULL,
  chart_type = NULL, custom_line_color_map = NULL, decimals = NULL,
  error = NULL, format = NULL, interpolate = NULL, legend = NULL,
  legend_target = NULL, linked = NULL, list = NULL, markers = NULL,
  max_data_size = NULL, mouseover = NULL, mousemove = NULL,
  mouseout = NULL, point_size = NULL, rollover_callback = NULL,
  show_confidence_band = NULL, show_rollover_text = NULL, target = NULL,
  transition_on_update = NULL, x_rug = NULL, y_rug = NULL)
}
\description{
Add graphic options to a MG plot object
}
\seealso{
\url{https://github.com/mozilla/metrics-graphics/wiki/Graphic}
}

