% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MG_layout.R
\name{MG_layout}
\alias{MG_layout}
\title{Add layout options to a MG plot object}
\usage{
MG_layout(buffer = NULL, buttom = NULL, height = NULL, left = NULL,
  right = NULL, small_height_threshold = NULL,
  small_width_threshold = NULL, top = NULL, width = NULL)
}
\arguments{
\item{buffer}{the buffer size in pixels between the actual chart area and the margins}

\item{buttom}{the size of the bottom margin in pixels}

\item{height}{the height of the entire graphic in pixels}

\item{left}{the size of the left margin in pixels}

\item{right}{the size of the right margin in pixels}

\item{small_height_threshold}{the height threshold in pixels for when the axis labels for a chart ought to be reduced in size}

\item{small_width_threshold}{the width threshold in pixels for when the axis labels for a chart ought to be reduced in size}

\item{top}{the size of the top margin in pixels}

\item{width}{the width of the entire graphic in pixels}
}
\description{
Add layout options to a MG plot object
}
\seealso{
\url{https://github.com/mozilla/metrics-graphics/wiki/Layout}
}

